requirements.txt
https://www.djangoproject.com/download/1.7c2/tarball/
Markdown==2.4.1
MySQL-python==1.2.5
Pillow==2.5.1
PyPDF2==1.22
argparse==1.2.1
diff-match-patch==20110725.1
dj-database-url==0.3.0
dj-static==0.0.6
django-easy-pdf==0.1.0
django-import-export==0.2.3
django-markupfield==1.2.0
django-ordered-model==0.4.0
django-recaptcha==1.0
django-toolbelt==0.0.1
gunicorn==18.0
html5lib==0.999
psycopg2==2.5.2
pycrypto==2.6.1
pystache==0.5.3
reportlab==2.7
six==1.7.3
static==1.0.2
static3==0.5.1
tablib==0.10.0
wsgiref==0.1.2
xhtml2pdf==0.0.6
django-downloadview==1.6
mock==1.0.1
requests==2.3.0
djrill==1.1.1
